Tests/LibWeb/Text/input/IntersectionObserver/implicit-root-target-visible-in-scroll-container.html

<!DOCTYPE html>
<style>
body { margin: 0; }
#scroller {
    width: 200px;
    height: 200px;
    overflow: hidden;
}
#target {
    width: 50px;
    height: 50px;
    background: red;
}
</style>
<script src="../include.js"></script>
<div id="scroller">
    <div id="target"></div>
</div>
<script>
    asyncTest(done => {
        // The target is at the top of a scroll container and visible.
        // With the implicit root, the target SHOULD be intersecting.
        let observer = new IntersectionObserver(entries => {
            entries.forEach(entry => {
                println(`isIntersecting=${entry.isIntersecting}, ratio=${entry.intersectionRatio}`);
                if (entry.isIntersecting) {
                    done();
                }
            });
        });

        observer.observe(document.getElementById("target"));
    });
</script>
