Tests/LibWeb/Text/expected/IntersectionObserver/implicit-root-target-visible-in-scroll-container.txt

isIntersecting=true, ratio=1
